split test module out to its own file + address other review comments
